Turned composite operator section into a module
